+++ KalikoSearch/Core/Collection.cs
@@ -3,7 +3,6 @@
     using System.IO;
     using System.Web;
     using Configuration;
-    using Lucene.Net.Analysis;
     using Lucene.Net.Store;
 
     public class Collection : IDisposable {
@@ -17,7 +16,7 @@
             var directory = GetDataDirectory();
             var analyzer = SearchAnalyzer.GetConfiguredAnalyzer();
             _searchIndexer = new SearchIndexer(directory, analyzer);
-            _searchFinder = new SearchFinder(_searchIndexer, analyzer);
+            _searchFinder = new SearchFinder(directory, analyzer);
         }
 
         public void AddDocument(IndexDocument document) {
@@ -44,8 +43,8 @@
             return _searchFinder.Search(queryString, searchFields, metaData, resultOffset, resultLength);
         }
 
-        public SearchResult FindSimular(string key, int resultOffset = 0, int resultLength = 10, bool matchCategory = true) {
-            return _searchFinder.FindSimular(key, resultOffset, resultLength, matchCategory);
+        public SearchResult FindSimular(string key, int resultOffset = 0, int resultLength = 10, bool matchCategory = true, string[] metaData = null) {
+            return _searchFinder.FindSimular(key, resultOffset, resultLength, matchCategory, metaData);
         }
 
         private FSDirectory GetDataDirectory() {