+++ KalikoSearch/Core/SearchFinder.cs
@@ -9,18 +9,21 @@
     using Lucene.Net.Search;
     using Lucene.Net.Search.Highlight;
     using Lucene.Net.Search.Similar;
+    using Lucene.Net.Store;
     using Version = Lucene.Net.Util.Version;
 
     public class SearchFinder : IDisposable {
         public const int MaxHits = 1000;
+        private readonly FSDirectory _directory;
         private readonly Analyzer _analyzer;
         private Searcher _searcher;
         private IndexReader _reader;
         private bool _open;
 
-        public SearchFinder(SearchIndexer indexer, Analyzer analyzer) {
+        public SearchFinder(FSDirectory directory, Analyzer analyzer) {
+            _directory = directory;
             _analyzer = analyzer;
-            _reader = indexer.IndexWriter.GetReader();
+            _reader = IndexReader.Open(directory, true);
             CreateIndexer();
             _open = true;
         }
@@ -50,15 +53,13 @@
         }
 
         public void Refresh() {
-            IndexReader newIndexReader = _reader.Reopen();
-            CloseReader();
             CloseSearcher();
 
-            _reader = newIndexReader;
+            _reader = IndexReader.Open(_directory, true);
             CreateIndexer();
         }
 
-        public SearchResult FindSimular(string key, int resultOffset, int resultLength, bool matchCategory) {
+        public SearchResult FindSimular(string key, int resultOffset, int resultLength, bool matchCategory, string[] metaData) {
             var pageQuery = new TermQuery(new Term("key", key));
             var topDocs = _searcher.Search(pageQuery, 1);
             if (topDocs.TotalHits == 0) {
@@ -119,12 +120,16 @@
                         PageId = pageId,
                         Path = document.Get("path"),
                         Title = document.Get("title"),
-                        Excerpt = document.Get("summary")
+                        Excerpt = document.Get("summary"),
+                        Summary = document.Get("summary"),
+                        Tags = document.Get("tags")
                     };
 
-                    //foreach (string key in metaData) {
-                    //    hit.MetaData.Add(key, document.Get(key));
-                    //}
+                    if (metaData != null) {
+                        foreach (var field in metaData) {
+                            hit.MetaData.Add(field, document.Get(field));
+                        }
+                    }
 
                     result.Hits.Add(hit);
                 }
@@ -168,7 +173,7 @@
             var result = new SearchResult {NumberOfHits = scoreDocs.Length};
 
             // Create highlighter
-            IFormatter formatter = new SimpleHTMLFormatter("<span class=\"search-highlight;\">", "</span>");
+            IFormatter formatter = new SimpleHTMLFormatter("<span class=\"search-highlight\">", "</span>");
             var fragmenter = new SimpleFragmenter(120);
             var scorer = new QueryScorer(query);
             var highlighter = new Highlighter(formatter, scorer) {TextFragmenter = fragmenter};
@@ -197,7 +202,9 @@
                         PageId = pageId,
                         Path = document.Get("path"),
                         Title = document.Get("title"), 
-                        Excerpt = excerpt
+                        Excerpt = excerpt,
+                        Summary = document.Get("summary"),
+                        Tags = document.Get("tags")
                     };