+++ ReleaseNotes.md
@@ -1,3 +1,12 @@
+### New in 3.0.3.5
+* Reduction of write locks
+* FindSimular category matchning improved
+* Possible to select which fields to search in 
+
+### New in 3.0.3.4
+* Fixed bug that disabled highlighting in search results
+* Updated reference to KalikoCMS.Core 1.0.0 to index when page published event is triggered
+
 ### New in 3.0.3.3
 * Fixed publishing date to use universal time zone instead of local to match the CMS internal format
 * Improved simular page search